Db_SIAP/requirements.txt
rpy2
pyodbc
threadpoolctl
asn1crypto==1.3.0
astroid==2.4.2
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
bleach==3.1.5
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
cryptography
cryptography-vectors==2.3.1
cycler==0.10.0
contextily==1.1.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.3
sqlalchemy==1.4.25
psycopg2==2.9.1
PyMySQL==1.0.2
pymssql==2.2.2
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
geopandas==0.9.0
imagesize==1.2.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keyring==21.2.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mapclassify==2.2.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.5
numpydoc==1.0.0
openpyxl==3.0.3
ordered-set==4.0.2
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyLaTeX==1.4.0
pylint==2.5.3
pyodbc 
sqlalchemy
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.16.0
pyshp==2.1.0
PySocks==1.7.1
python-dateutil==2.7.2
pytz==2020.1
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.1
python-dateutil==2.7.2
pycrypto==2.6.1
QtAwesome==0.7.2
qtconsole==4.7.5
QtPy==1.9.0
reportlab==3.5.59
requests==2.18.4
rope==0.17.0
scikit-learn==0.23.1
scipy==1.5.3
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==3.3.6
spyder-kernels==0.5.2
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
xlwt
zipp==3.1.0